Add backup service

+++ db_creation.py
@@ -1,9 +1,19 @@
+"""
+This file connects to the trees database running locally and
+resets the tables in the database using the data in tree_history.csv and tree_info.csv
+
+It should only be run when first initializing the server when there is no existing data, or
+to recover from a failure of the postgres database.
+"""
+
+import os.path
 import pandas as pd
-from config import *
+from passlib.hash import pbkdf2_sha256
 import psycopg2
 from schemas.table_schemas import TreeInfo, TreeHistory, Users
 from sqlmodel import SQLModel, create_engine, Session
-from passlib.hash import pbkdf2_sha256
+
+from config import *
 
 
 def main():
@@ -12,30 +22,24 @@ def main():
     while user_input != "tree":
         user_input = input("ERROR: You did not enter \"tree\" try again.\n")
 
+    # Ensures that csvs to pull from actually exist
+    if not (os.path.isfile("./active_csvs/tree_history.csv") and os.path.isfile("./active_csvs/tree_info.csv")):
+        raise Exception("tree_history.csv and/or tree_info.csv do not exist in the active_csvs folder.\nProgram cannot run without those EXACT files existing in that EXACT location./")
+
+    # Opens and reads db csv files
     tree_history = pd.read_csv("./active_csvs/tree_history.csv")
     tree_info = pd.read_csv("./active_csvs/tree_info.csv")
 
     # Tries to connect to postgres database on default port if possible
     default_user = Users(username=ADMIN_USERNAME, email=ADMIN_EMAIL, full_name=ADMIN_NAME, hashed_password=pbkdf2_sha256.hash(ADMIN_PASSWORD), data_permissions=True, user_permissions=True)
-    try:
-        engine = create_engine(f"postgresql://{LOCAL_DB_USER}:{LOCAL_DB_PASS}@localhost:5432/{LOCAL_DB_NAME}", echo=True)
-        SQLModel.metadata.drop_all(engine)
-        SQLModel.metadata.create_all(engine)
-        tree_info.to_sql('treeinfo', engine, if_exists="append", index=False)
-        tree_history.to_sql('treehistory', engine, if_exists="append", index=False)
-        with Session(engine) as session:
-            session.add(default_user)
-            session.commit()
-
-    except Exception as e:
-        engine = create_engine(f"postgresql://{LOCAL_DB_USER}:{LOCAL_DB_PASS}@localhost:5433/{LOCAL_DB_NAME}", echo=True)
-        SQLModel.metadata.drop_all(engine)
-        SQLModel.metadata.create_all(engine)
-        tree_info.to_sql('treeinfo', engine, if_exists="append", index=False)
-        tree_history.to_sql('treehistory', engine, if_exists="append", index=False)
-        with Session(engine) as session:
-            session.add(default_user)
-            session.commit()
+    engine = create_engine(f"postgresql://{LOCAL_DB_USER}:{LOCAL_DB_PASS}@localhost:{LOCAL_DB_PORT}/{LOCAL_DB_NAME}", echo=True)
+    SQLModel.metadata.drop_all(engine)
+    SQLModel.metadata.create_all(engine)
+    tree_info.to_sql('treeinfo', engine, if_exists="append", index=False)
+    tree_history.to_sql('treehistory', engine, if_exists="append", index=False)
+    with Session(engine) as session:
+        session.add(default_user)
+        session.commit()
 
 
 if __name__ == "__main__":
